Share the Waypoint and NpcNavigator buses and the Waypoint configuration
